requirements.txt
Flask~=3.1.0
Flask-Cors~=5.0.0
python-dotenv~=1.0.1
langchain-community~=0.3.14
langchain~=0.3.14
langchain-google-genai~=2.0.7
langchain-core~=0.3.29
langchain-openai~=0.2.14
requests~=2.32.3
pdf2image~=1.17.0